requirements.txt
# 영어 발음 분석 파이프라인 - Requirements
# Python 3.10.12 환경에서 테스트됨
# 실제 사용되는 라이브러리만 포함

# ==== 웹 프레임워크 & API ====
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# ==== 데이터 검증 및 스키마 ====
pydantic==2.11.5
pydantic-core==2.33.2

# ==== HTTP 클라이언트 & 비동기 처리 ====
httpx==0.25.2
aiohttp==3.12.7
aiosignal==1.3.2
async-timeout==5.0.1

# ==== 데이터베이스 ====
# MongoDB (비동기)
motor==3.3.2
pymongo==4.6.0

# MariaDB/MySQL
pymysql==1.1.0

# ==== AWS SDK ====
boto3==1.34.34
botocore==1.34.34

# ==== OpenAI & GPT ====
openai==1.12.0

# ==== 음성 처리 & 분석 ====
# Montreal Forced Aligner는 conda에서 설치
# conda install -c conda-forge montreal-forced-aligner

# Praat & 음성 분석
praat-parselmouth==0.4.5
praat-textgrids==1.4.0

# 오디오 라이브러리  
librosa>=0.10.0
soundfile
audioread

# WhisperX 음성 인식
whisperx==3.3.4

# PyAnnote 화자 분리
pyannote.audio==3.3.2

# ==== 수치 계산 & 데이터 처리 ====
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
statistics  # built-in
math  # built-in

# ==== 자연어 처리 ====
spacy>=3.7.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl

# 구문 분석
benepar==0.2.0

# 텍스트 그리드 처리
textgrids

# ==== 시각화 (선택사항) ====
matplotlib>=3.7.0

# ==== 기타 유틸리티 ====
# 정규 표현식, OS, 시스템 모듈들은 built-in
# re, os, sys, json, csv, glob, subprocess, tempfile, shutil, pathlib
# random, string, collections, datetime, typing, logging, asyncio, signal

# 파일 처리
wave  # built-in
contextlib  # built-in

# ==== YAML 파일 처리 ====
PyYAML==6.0.1

# ==== 개발 도구 (선택사항) ====
# argparse  # built-in 